pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyiceberg_firestore_gcs"
version = "0.3.1"
description = "Opteryx Cloud Iceberg Catalog"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-cloud-firestore==2.22.0",
    "google-cloud-storage==3.7.0",
    "orso==0.0.226",
    "opteryx-core",
    "pyarrow==22.0.0",
    "pyiceberg==0.10.0",
    "requests==2.32.5"
]

[tool.setuptools.package-dir]
"" = "."

[tool.setuptools.packages.find]
where = ["."]

[tool.isort]
profile = "black"
extend_skip_glob = ["tests/**"]
skip_gitignore = true
line_length = 100
multi_line_output = 9
force_single_line = true
float_to_top = true
ensure_newline_before_comments = true

[tool.ruff]
line-length = 100
indent-width = 4
target-version = 'py313'

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 100
